#51 #53 Fix lint errors and warnings

+++ mmr-backend/api.py
@@ -1,4 +1,4 @@
-from ctypes.util import find_library
+from typing import Iterable, Optional
 from book import Book
 from library import Library
 import os
@@ -21,66 +21,72 @@ mmr = FastAPI()
 
 
 @mmr.get("/books")
-async def get_books():
+async def get_books() -> list[dict]:
     return list(map(lambda book: Book.to_dict(book), book_list))
 
 
 @mmr.get("/books/{isbn}")
-async def get_book(isbn: str):
-    return list(filter(lambda book: book.to_dict()["ISBN"] == isbn, book_list))
+async def get_book(isbn: str) -> list[dict]:
+    matching_book: Iterable = filter(lambda book: book.to_dict()[
+        "ISBN"] == isbn, book_list)
+    return list(matching_book)
 
 
 @mmr.get("/libraries/{user}")
-async def get_libraries(user: str):
+async def get_libraries(user: str) -> list[dict]:
     libraries_from_user = filter(
         lambda lib: lib.get_owner() == user, libraries)
     return list(map(lambda lib: lib.to_dict(), libraries_from_user))
 
 
-@mmr.get("/libraries/{user}/{library}")
-async def get_library(user: str, library: str):
-    library = find_library(user, library)
-    return library.to_dict()
+@mmr.get("/libraries/{user}/{library_name}")
+async def get_library(user: str, library_name: str) -> dict:
+    library: Optional[Library] = find_library(user, library_name)
+    return library.to_dict() if library else {}
 
 
-@mmr.delete("/libraries/{user}/{library}")
-async def delete_library(user: str, library: str):
-    library = find_library(user, library)
+@mmr.delete("/libraries/{user}/{library_name}")
+async def delete_library(user: str, library_name: str) -> None:
+    library: Optional[Library] = find_library(user, library_name)
     if library:
         libraries.remove(library)
 
 
-@mmr.put("/libraries/{user}/{library}/{new_name}")
-async def update_library_name(user: str, library: str, new_name: str):
-    library = find_library(user, library)
-    library.set_name(new_name)
+@mmr.put("/libraries/{user}/{library_name}/{new_name}")
+async def update_library_name(user: str, library_name: str, new_name: str) -> None:
+    library: Optional[Library] = find_library(user, library_name)
+    if library:
+        library.set_name(new_name)
 
 
-@mmr.post("/libraries/{user}/{library}")
-async def create_library(user: str, library: str):
-    libraries.append(Library(user, library, list()))
+@mmr.post("/libraries/{user}/{library_name}")
+async def create_library(user: str, library_name: str) -> None:
+    libraries.append(Library(user, library_name, list()))
 
 
-@mmr.post("/libraries/{user}/{library}/{isbn}")
-async def add_library_entry(user: str, library: str, isbn: str):
-    library = find_library(user, library)
-    library.add_entry(Library.Entry(isbn))
+@mmr.post("/libraries/{user}/{library_name}/{isbn}")
+async def add_library_entry(user: str, library_name: str, isbn: str) -> None:
+    library: Optional[Library] = find_library(user, library_name)
+    if library:
+        library.add_entry(Library.Entry(isbn))
 
 
-@mmr.delete("/libraries/{user}/{library}/{isbn}")
-async def add_library_entry(user: str, library: str, isbn: str):
-    library = find_library(user, library)
-    library.remove_entry(isbn)
+@mmr.delete("/libraries/{user}/{library_name}/{isbn}")
+async def remove_library_entry(user: str, library_name: str, isbn: str) -> None:
+    library: Optional[Library] = find_library(user, library_name)
+    if library:
+        library.remove_entry(isbn)
 
 
-@mmr.put("/libraries/{user}/{library}/{isbn}/{score}/{status}")
-async def add_library_entry(user: str, library: str, isbn: str, score: int, status: Library.ReadingStatus):
-    library = find_library(user, library)
-    library.update_entry(isbn, Library.Entry(isbn, score, status))
+@mmr.put("/libraries/{user}/{library_name}/{isbn}/{score}/{status}")
+async def update_library_entry(user: str, library_name: str, isbn: str, score: int, status: Library.ReadingStatus) -> None:
+    library: Optional[Library] = find_library(user, library_name)
+    if library:
+        library.update_entry(isbn, Library.Entry(isbn, score, status))
 
 
-def find_library(owner: str, name: str) -> Library:
+def find_library(owner: str, name: str) -> Optional[Library]:
     occurrencies = list(filter(
         lambda lib: lib.get_owner() == owner and lib.get_name() == name, libraries))
 
-    return occurrencies[0] if len(occurrencies) else {}
+    return occurrencies[0] if len(occurrencies) else None
